MINT-1701. Refactor time-locked wallet form. Fix errors.
