+++ src/components/wallet/AddWalletWidget/SelectEthWallet/SelectEthWallet.jsx
@@ -63,6 +63,7 @@ export default class SelectEthWallet extends PureComponent {
         type: '2FA',
         icon: 'security-circle',
         description: `${prefix}.fa.description`,
+        disabled: true,
       },
       {
         title: `${prefix}.ms.title`,
