+++ src/components/wallet/OwnersList/OwnersList.jsx
@@ -24,7 +24,7 @@ function mapStateToProps (state, props) {
   const newOwner = selector(state, FIELD_NEW_ADDRESS)
 
   return {
-    isDisabled: !!errors[ FIELD_NEW_ADDRESS ] || !newOwner,
+    isDisabled: !!errors[FIELD_NEW_ADDRESS] || !newOwner,
     newOwner,
     account: state.get(DUCK_SESSION).account,
   }
@@ -86,8 +86,7 @@ export default class OwnersList extends PureComponent {
           <div styleName='addOwnerField'>
             <Field
               component={TextField}
-              hintText={<Translate value={`${prefix}.floatText`} />}
-              hintStyle={globalStyles.textField.hintStyle}
+              label={<Translate value={`${prefix}.floatText`} />}
               name={FIELD_NEW_ADDRESS}
               validate={validate}
               fullWidth
